Separate out evaluator.
