+++ src/Juvix/Core/IR.hs
@@ -1,8 +1,10 @@
 module Juvix.Core.IR
   ( module Juvix.Core.IR.Types,
     module Juvix.Core.IR.Typechecker,
+    module Juvix.Core.IR.Evaluator,
   )
 where
 
+import Juvix.Core.IR.Evaluator
 import Juvix.Core.IR.Typechecker
 import Juvix.Core.IR.Types
